// src/gen/api/types/ListCommentReplies.ts
/**
 * Generated by Kubb (https://kubb.dev/).
 * Do not edit manually.
 */

import type { Comment } from './Comment.ts'

export type ListCommentRepliesPathParams = {
  /**
   * @description UUID identifier
   * @type string, uuid
   */
  comment_id: string
}

export type ListCommentRepliesQueryParams = {
  /**
   * @type string | undefined
   */
  cursor?: string
  /**
   * @minLength 1
   * @maxLength 50
   * @default 20
   * @type integer | undefined
   */
  limit?: number
}

/**
 * @description Cursor-paginated replies
 */
export type ListCommentReplies200 = {
  /**
   * @type array
   */
  data: Comment[]
  /**
   * @type object
   */
  pagination: {
    /**
     * @type string,null
     */
    nextCursor: string | null
    /**
     * @type boolean
     */
    hasNext: boolean
    /**
     * @type number
     */
    limit: number
  }
}

/**
 * @description Validation failed
 */
export type ListCommentReplies400 = {
  /**
   * @type string
   */
  error: string
  /**
   * @type string
   */
  message: string
  /**
   * @type string
   */
  code: string
  /**
   * @type string | undefined
   */
  requestId?: string
  /**
   * @type object | undefined
   */
  details?: {
    [key: string]: string
  }
}

/**
 * @description Missing or invalid access token
 */
export type ListCommentReplies401 = {
  /**
   * @type string
   */
  error: string
  /**
   * @type string
   */
  message: string
  /**
   * @type string
   */
  code: string
  /**
   * @type string | undefined
   */
  requestId?: string
  /**
   * @type object | undefined
   */
  details?: {
    [key: string]: string
  }
}

/**
 * @description Not found
 */
export type ListCommentReplies404 = {
  /**
   * @type string
   */
  error: string
  /**
   * @type string
   */
  message: string
  /**
   * @type string
   */
  code: string
  /**
   * @type string | undefined
   */
  requestId?: string
  /**
   * @type object | undefined
   */
  details?: {
    [key: string]: string
  }
}

/**
 * @description Internal server error
 */
export type ListCommentReplies500 = {
  /**
   * @type string
   */
  error: string
  /**
   * @type string
   */
  message: string
  /**
   * @type string
   */
  code: string
  /**
   * @type string | undefined
   */
  requestId?: string
  /**
   * @type object | undefined
   */
  details?: {
    [key: string]: string
  }
}

export type ListCommentRepliesQueryResponse = ListCommentReplies200

export type ListCommentRepliesQuery = {
  Response: ListCommentReplies200
  PathParams: ListCommentRepliesPathParams
  QueryParams: ListCommentRepliesQueryParams
  Errors:
    | ListCommentReplies400
    | ListCommentReplies401
    | ListCommentReplies404
    | ListCommentReplies500
}